Move LayoutContext interface to Layout
